requirements.txt
jwt
pyjwt
Django==4.0.4
django-cors-headers==3.11.0
django-rest-swagger==2.2.0
django-sslserver==0.22
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
psycopg2
docxtpl
docx
boto3


